Plot otuput of the overlap_df function as a heatmap.

Plot otuput of the overlap_df function as a heatmap.

Usage

corr_heatmap(rec, steps = steps_ids(rec, "da"), font_size = 15)

# S4 method for recipe
corr_heatmap(rec, steps = steps_ids(rec, "da"), font_size = 15)

# S4 method for prep_recipe
corr_heatmap(rec, steps = steps_ids(rec, "da"), font_size = 15)

Arguments

rec

A recipe object.

steps

Character vector with step_ids to take in account.

font_size

Size of the axis font.

Value

heatmap

Examples

data(test_prep_rec)

## Running the function returns a UpSet plot ordered by frequency.
corr_heatmap(test_prep_rec)

## If you want to exclude a method for the plot, you can remove it with the
## step parameter. In the following example we eliminate from the graph the
## results of maaslin
corr_heatmap(test_prep_rec, steps = steps_ids(test_prep_rec, "da")[-1])

## intersection_plt function needs a prep-recipe. If you pass a a non-prep
## recipe the output is an error.
data(test_rec)
if (FALSE) df <- corr_heatmap(test_rec)